Move ChromaSampling to v_frame
