+++ src/api/color.rs
@@ -12,8 +12,6 @@ use crate::serialize::*;
 use arg_enum_proc_macro::ArgEnum;
 use num_derive::FromPrimitive;
 
-use std::fmt;
-
 /// Sample position for subsampled chroma
 #[derive(
   Copy, Clone, Debug, PartialEq, FromPrimitive, Serialize, Deserialize,
@@ -36,71 +34,7 @@ impl Default for ChromaSamplePosition {
   }
 }
 
-/// Chroma subsampling format
-#[derive(
-  Copy, Clone, Debug, PartialEq, FromPrimitive, Serialize, Deserialize,
-)]
-#[repr(C)]
-pub enum ChromaSampling {
-  /// Both vertically and horizontally subsampled.
-  Cs420,
-  /// Horizontally subsampled.
-  Cs422,
-  /// Not subsampled.
-  Cs444,
-  /// Monochrome.
-  Cs400,
-}
-
-impl fmt::Display for ChromaSampling {
-  fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
-    write!(
-      f,
-      "{}",
-      match self {
-        ChromaSampling::Cs420 => "4:2:0",
-        ChromaSampling::Cs422 => "4:2:2",
-        ChromaSampling::Cs444 => "4:4:4",
-        ChromaSampling::Cs400 => "Monochrome",
-      }
-    )
-  }
-}
-
-impl Default for ChromaSampling {
-  fn default() -> Self {
-    ChromaSampling::Cs420
-  }
-}
-
-impl ChromaSampling {
-  /// Provides the amount to right shift the luma plane dimensions to get the
-  ///  chroma plane dimensions.
-  /// Only values 0 or 1 are ever returned.
-  /// The plane dimensions must also be rounded up to accommodate odd luma plane
-  ///  sizes.
-  /// Cs400 returns None, as there are no chroma planes.
-  pub fn get_decimation(self) -> Option<(usize, usize)> {
-    use self::ChromaSampling::*;
-    match self {
-      Cs420 => Some((1, 1)),
-      Cs422 => Some((1, 0)),
-      Cs444 => Some((0, 0)),
-      Cs400 => None,
-    }
-  }
-
-  /// Calculates the size of a chroma plane for this sampling type, given the luma plane dimensions.
-  pub fn get_chroma_dimensions(
-    self, luma_width: usize, luma_height: usize,
-  ) -> (usize, usize) {
-    if let Some((ss_x, ss_y)) = self.get_decimation() {
-      ((luma_width + ss_x) >> ss_x, (luma_height + ss_y) >> ss_y)
-    } else {
-      (0, 0)
-    }
-  }
-}
+pub use v_frame::pixel::ChromaSampling;
 
 /// Supported Color Primaries
 ///
