+++ v_frame/Cargo.toml
@@ -6,7 +6,12 @@ license = "BSD-2-Clause"
 authors = ["Luca Barbato <lu_zero@gentoo.org>"]
 edition = "2018"
 
+[features]
+serialize = ["serde"]
 
 [dependencies]
 cfg-if = "0.1"
 num-traits = "0.2"
+num-derive = "0.3"
+noop_proc_macro = "0.2.0"
+serde = { version = "1.0", features = ["derive"], optional = true }
