+++ v_frame/src/pixel.rs
@@ -7,7 +7,11 @@
 // Media Patent License 1.0 was not distributed with this source code in the
 // PATENTS file, you can obtain it at www.aomedia.org/license/patent.
 
+use crate::serialize::{Deserialize, Serialize};
+
+use num_derive::FromPrimitive;
 use num_traits::{AsPrimitive, PrimInt, Signed};
+use std::fmt;
 use std::fmt::{Debug, Display};
 use std::mem::size_of;
 use std::ops::AddAssign;
@@ -139,4 +143,70 @@ impl Coefficient for i16 {
 }
 impl Coefficient for i32 {
   type Pixel = u16;
+}
+
+/// Chroma subsampling format
+#[derive(
+  Copy, Clone, Debug, PartialEq, FromPrimitive, Serialize, Deserialize,
+)]
+#[repr(C)]
+pub enum ChromaSampling {
+  /// Both vertically and horizontally subsampled.
+  Cs420,
+  /// Horizontally subsampled.
+  Cs422,
+  /// Not subsampled.
+  Cs444,
+  /// Monochrome.
+  Cs400,
+}
+
+impl fmt::Display for ChromaSampling {
+  fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
+    write!(
+      f,
+      "{}",
+      match self {
+        ChromaSampling::Cs420 => "4:2:0",
+        ChromaSampling::Cs422 => "4:2:2",
+        ChromaSampling::Cs444 => "4:4:4",
+        ChromaSampling::Cs400 => "Monochrome",
+      }
+    )
+  }
+}
+
+impl Default for ChromaSampling {
+  fn default() -> Self {
+    ChromaSampling::Cs420
+  }
+}
+
+impl ChromaSampling {
+  /// Provides the amount to right shift the luma plane dimensions to get the
+  ///  chroma plane dimensions.
+  /// Only values 0 or 1 are ever returned.
+  /// The plane dimensions must also be rounded up to accommodate odd luma plane
+  ///  sizes.
+  /// Cs400 returns None, as there are no chroma planes.
+  pub fn get_decimation(self) -> Option<(usize, usize)> {
+    use self::ChromaSampling::*;
+    match self {
+      Cs420 => Some((1, 1)),
+      Cs422 => Some((1, 0)),
+      Cs444 => Some((0, 0)),
+      Cs400 => None,
+    }
+  }
+
+  /// Calculates the size of a chroma plane for this sampling type, given the luma plane dimensions.
+  pub fn get_chroma_dimensions(
+    self, luma_width: usize, luma_height: usize,
+  ) -> (usize, usize) {
+    if let Some((ss_x, ss_y)) = self.get_decimation() {
+      ((luma_width + ss_x) >> ss_x, (luma_height + ss_y) >> ss_y)
+    } else {
+      (0, 0)
+    }
+  }
 }
